update at march 27,2018

+++ DirectAnalysis/perl/SumInputFilesScripts.pl
@@ -4,9 +4,12 @@ chomp($workdir =`pwd -P |sed 's\\perl\\\\g '`);
 print "Printed: Work Dir. = ".$workdir."\n\n";
 
 $datapath  = "/eos/ams/group/dbar/release_v4/e1_vdev_180213/neg/ISS.B950/pass6";
-$mcP_path  = "/eos/ams/group/dbar/release_v4/e1_vdev_180213/full/Pr.B1082/pr.pl1.05100.2_01";
+$mcP_path  = "/eos/ams/group/dbar/release_v4/e1_vdev_180213/full/Pr.B1082/pr.pl1.l1.05100.2_01/";
 $mcD_path  = "/eos/ams/group/dbar/release_v4/e1_vdev_180213/full/D.B1081/d.pl1.l1.0_5200.2_01";
 $mcHe_path = "/eos/ams/group/dbar/release_v4/e1_vdev_180213/full/He.B1081/he4.pl1.l1.2200.2_01";
+$mcT_path  = "/eos/ams/group/dbar/release_v4/e1_vdev_180213/full/T.B1059/t.pl1.0_520_GG_BlicDPMJet/";
+
+
 $out_path  = "/eos/ams/user/f/fdimicco/";
 
 #$ntuplepath  = "/eos/ams/group/dbar/TrentoNTuples/$ARGV[0]-$ARGV[1]";
@@ -15,11 +18,16 @@ $ntuplepath  = "/eos/ams/user/f/fdimicco/AnalysisNTuples/$ARGV[0]-$ARGV[1]";
 
 
 #use warnings;
-system("rm $workdir/InputFileLists/*");
+if($ARGV[3]==0) { system("rm $workdir/InputFileLists/*"); }
+else { system("rm $workdir/InputNtupleLists/*");}
+
 
 print "Listing All Data Files..\n";
 chomp (@Rootuple = `ls  $datapath | grep -v "log" |  sed s/.root//g`);
 $num_Rootuple = scalar(@Rootuple);
+
+chomp (@NTuple = `ls  $ntuplepath | grep -v "log" |  sed s/.root//g`);
+$num_NTuple = scalar(@NTuple);
 
 print "Total Files: ".$num_Rootuple."\n";
 @rootuple;
@@ -39,11 +47,14 @@ print "Files in the requested period: ".$num_rootuple."\n";
 
 for ($n=0;$n<$njobs; $n++)
 {
-	open(OUT,">","$workdir/InputFileLists/FileListDT$n.txt");
 	if($ARGV[3]==1){
-		print OUT  "$ntuplepath/Result$n.root\n";
+		open(OUT,">","$workdir/InputNtupleLists/FileListDT$n.txt");
+		for ($j=($num_NTuple)/$njobs*$n ; $j<($num_NTuple)/$njobs*($n+1) ; $j++){
+			print OUT  "$ntuplepath/$NTuple[$j].root\n";
+		}
 	}
 	else{
+		open(OUT,">","$workdir/InputFileLists/FileListDT$n.txt");
 		for ($j=($num_rootuple)/$njobs*$n ; $j<($num_rootuple)/$njobs*($n+1) ; $j++)
 		{
 			print OUT  "$datapath/$rootuple[$j].root\n";
@@ -70,16 +81,25 @@ $num_MC_He = scalar(@MC_He);
 print "Total Files MC He: ".$num_MC_He."\n";
 
 
+chomp (@MC_T = `ls $mcT_path  | grep -v "log" |  sed s/.root//g`);
+$num_MC_T = scalar(@MC_T);
+
+
+print "Total Files MC T: ".$num_MC_T."\n";
+
+
 
 for ($n=0;$n<$njobs; $n++)
 {
 
-        open(OUT,">","$workdir/InputFileLists/FileListMC$n.txt");
-
 	if($ARGV[3]==1){
-                print OUT  "$ntuplepath/Result$n.root\n";
-        }
+		open(OUT,">","$workdir/InputNtupleLists/FileListMC$n.txt");
+		for ($j=($num_NTuple)/$njobs*$n ; $j<($num_NTuple)/$njobs*($n+1) ; $j++){
+			print OUT  "$ntuplepath/$NTuple[$j].root\n";
+		}
+	}
 	else{
+		open(OUT,">","$workdir/InputFileLists/FileListMC$n.txt");
 
 		for ($j=($num_MC_P)/$njobs*$n ; $j<($num_MC_P)/$njobs*($n+1) ; $j++)
 		{
@@ -93,6 +113,12 @@ for ($n=0;$n<$njobs; $n++)
 		{
 			print OUT  "$mcHe_path/$MC_He[$j].root\n"
 		}
+		for ($j=($num_MC_T)/$njobs*$n ; $j<($num_MC_T)/$njobs*($n+1) ; $j++)
+		{
+			print OUT  "$mcT_path/$MC_T[$j].root\n"
+		}
+
+
 	}
 }
